requirements.txt
langchain
langchain_groq
python-dotenv
pydantic
requests
django